physical_ai/main_server/requirements_hub.txt
eventlet
flask
flask-socketio
psycopg2-binary
python-dotenv
python-socketio